Add files via upload
I have done as part of my role for a group assignment completed the critical Create, Read Update and Delete (CRUD) operations which are a necessity for the creation, edition and deletion of RSS feeds for an web-based application Tailor Edited News (TEN). The Tailor Edited News ten will be responsive across a devices. Devices include; A Personal Computer (PC), An Computer Tablet and Smartphone(s).